application/facedetect/experiment/requirements.txt
aiohttp==3.6.2
asyncio==3.4.3
pandas==1.0.1
pyyaml==6.0